+++ Cargo.toml
@@ -22,6 +22,16 @@ members = [
 	"website/other/bezier-rs-demos/wasm",
 ]
 exclude = ["node-graph/gpu-compiler"]
+default-members = [
+	"editor",
+	"frontend/wasm",
+	"node-graph/gcore",
+	"node-graph/gstd",
+	"node-graph/graph-craft",
+	"node-graph/graphene-cli",
+	"node-graph/interpreted-executor",
+	"node-graph/node-macro",
+]
 resolver = "2"
 
 [workspace.dependencies]
