+++ editor/Cargo.toml
@@ -20,14 +20,17 @@ gpu = [
 	"wgpu-executor",
 	"gpu-executor",
 ]
+tauri = ["ron", "decouple-execution"]
+decouple-execution = []
 resvg = ["graphene-std/resvg"]
 vello = ["graphene-std/vello", "resvg", "graphene-core/vello"]
+ron = ["dep:ron"]
 
 [dependencies]
 # Local dependencies
 graphite-proc-macros = { path = "../proc-macros" }
 graph-craft = { path = "../node-graph/graph-craft" }
-interpreted-executor = { path = "../node-graph/interpreted-executor" }
+interpreted-executor = { path = "../node-graph/interpreted-executor", features = ["serde"] }
 graphene-core = { path = "../node-graph/gcore" }
 graphene-std = { path = "../node-graph/gstd", features = ["serde"] }
 
@@ -69,6 +72,7 @@ gpu-executor = { path = "../node-graph/gpu-executor", optional = true }
 # Optional workspace dependencies
 wasm-bindgen = { workspace = true, optional = true }
 wasm-bindgen-futures = { workspace = true, optional = true }
+ron = { workspace = true, optional = true }
 
 [dev-dependencies]
 # Workspace dependencies
