+++ editor/src/messages/tool/transform_layer/transform_layer_message_handler.rs
@@ -713,15 +713,10 @@ impl MessageHandler<TransformLayerMessage, TransformData<'_>> for TransformLayer
 
 #[cfg(test)]
 mod test_transform_layer {
+	use crate::messages::portfolio::document::graph_operation::{transform_utils, utility_types::ModifyInputsContext};
 	use crate::messages::portfolio::document::utility_types::misc::GroupFolderType;
-	use crate::messages::{
-		portfolio::document::graph_operation::{
-			transform_utils,
-			utility_types::{ModifyInputsContext, TransformIn},
-		},
-		prelude::Message,
-		tool::transform_layer::transform_layer_message_handler::VectorModificationType,
-	};
+	use crate::messages::prelude::Message;
+	use crate::messages::tool::transform_layer::transform_layer_message_handler::VectorModificationType;
 	use crate::test_utils::test_prelude::*;
 	use glam::DAffine2;
 	use graphene_core::vector::PointId;
@@ -1232,7 +1227,7 @@ mod test_transform_layer {
 
 		// Test 4: Transform layers inside transformed group
 		let child_layer_id = {
-			let mut document = editor.active_document_mut();
+			let document = editor.active_document_mut();
 			let group_children = document.network_interface.downstream_layers(&group_layer.to_node(), &[]);
 			if !group_children.is_empty() {
 				Some(LayerNodeIdentifier::new(group_children[0], &document.network_interface, &[]))
