+++ frontend/package.json
@@ -19,6 +19,8 @@
 		"lint-fix": "eslint . --fix && tsc --noEmit",
 		"---------- INTERNAL ----------": "",
 		"setup": "node package-installer.js",
+		"tauri:dev": "vite",
+		"tauri:build": "wasm-pack build ./wasm --target=web --features=tauri",
 		"wasm:build-dev": "wasm-pack build ./wasm --dev --target=web",
 		"wasm:build-profiling": "wasm-pack build ./wasm --profiling --target=web",
 		"wasm:build-production": "wasm-pack build ./wasm --release --target=web",
@@ -27,7 +29,6 @@
 		"wasm:watch-production": "cargo watch --postpone --watch-when-idle --workdir=wasm --shell \"wasm-pack build . --release --target=web -- --color=always\""
 	},
 	"dependencies": {
-		"@tauri-apps/api": "^1.6.0",
 		"class-transformer": "^0.5.1",
 		"idb-keyval": "^6.2.1",
 		"reflect-metadata": "^0.2.2"
