+++ frontend/src-tauri/capabilities/desktop.json
@@ -1,0 +1,6 @@
+{
+	"identifier": "desktop-capability",
+	"platforms": ["macOS", "windows", "linux"],
+	"windows": ["main"],
+	"permissions": ["http:default"]
+}
