+++ frontend/src-tauri/capabilities/migrated.json
@@ -1,0 +1,49 @@
+{
+	"identifier": "migrated",
+	"description": "permissions that were migrated from v1",
+	"local": true,
+	"windows": ["main"],
+	"permissions": [
+		"core:default",
+		"core:window:allow-create",
+		"core:window:allow-center",
+		"core:window:allow-request-user-attention",
+		"core:window:allow-set-resizable",
+		"core:window:allow-set-maximizable",
+		"core:window:allow-set-minimizable",
+		"core:window:allow-set-closable",
+		"core:window:allow-set-title",
+		"core:window:allow-maximize",
+		"core:window:allow-unmaximize",
+		"core:window:allow-minimize",
+		"core:window:allow-unminimize",
+		"core:window:allow-show",
+		"core:window:allow-hide",
+		"core:window:allow-close",
+		"core:window:allow-set-decorations",
+		"core:window:allow-set-always-on-top",
+		"core:window:allow-set-content-protected",
+		"core:window:allow-set-size",
+		"core:window:allow-set-min-size",
+		"core:window:allow-set-max-size",
+		"core:window:allow-set-position",
+		"core:window:allow-set-fullscreen",
+		"core:window:allow-set-focus",
+		"core:window:allow-set-icon",
+		"core:window:allow-set-skip-taskbar",
+		"core:window:allow-set-cursor-grab",
+		"core:window:allow-set-cursor-visible",
+		"core:window:allow-set-cursor-icon",
+		"core:window:allow-set-cursor-position",
+		"core:window:allow-set-ignore-cursor-events",
+		"core:window:allow-start-dragging",
+		"core:webview:allow-print",
+		"shell:allow-execute",
+		"shell:allow-open",
+		"http:default",
+		"core:app:allow-app-show",
+		"core:app:allow-app-hide",
+		"shell:default",
+		"http:default"
+	]
+}
